+++ src/main/java/dao/CodecoolerDAO.java
@@ -1,6 +1,8 @@
 package dao;
 
 import model.items.Artifact;
+import model.users.Codecooler;
+import model.users.User;
 import org.postgresql.core.SqlCommand;
 
 import java.sql.Connection;
@@ -12,6 +14,119 @@ import java.util.List;
 
 
 public class CodecoolerDAO implements ICodecoolerDAO {
+    Connection connection;
+    DBCreator dbCreator;
+
+    public CodecoolerDAO() {
+        dbCreator = new DBCreator();
+
+    }
+
+    public void createCodecooler(User user, Codecooler codecooler) {
+        createUser(user);
+        try {
+            int userID = getUserIdWithLogin(user);
+            codecooler.setId(userID);
+            createStudent(codecooler);
+        } catch(DBException e) {
+            e.printStackTrace();
+            System.out.println("Failed to fetch user with this login");
+        }
+
+    }
+
+    private void createUser(User user){
+        String query = "INSERT INTO users(login, password, usertype) VALUES (?,?,?)";
+        PreparedStatement statement = null;
+
+        try {
+
+            connection = dbCreator.connectToDatabase();
+            statement = connection.prepareStatement(query);
+
+
+            statement.setString(1, user.getLogin());
+            statement.setString(2, user.getPassword());
+            statement.setString(3, user.getUserType());
+
+
+            statement.executeUpdate();
+            connection.close();
+
+        } catch (SQLException e){
+            e.printStackTrace();
+        }
+
+    }
+
+    private void createStudent(Codecooler codecooler){
+        String query = "INSERT INTO studentpersonals(address, class_id, coolcoins, email, experience_points," +
+                " first_name, last_name, phone_number, user_id) VALUES (?,?,?,?,?,?,?,?,?)";
+        PreparedStatement statement = null;
+
+        try {
+            connection = dbCreator.connectToDatabase();
+
+            statement = connection.prepareStatement(query);
+
+
+            statement.setString(1, codecooler.getAdress());
+            statement.setInt(2, codecooler.getRoomID());
+            statement.setInt(3, codecooler.getAmmountOfCoins());
+            statement.setString(4, codecooler.getEmail());
+            statement.setInt(5, codecooler.getLvlOfExp());
+            statement.setString(6, codecooler.getFirstName());
+            statement.setString(7, codecooler.getLastName());
+            statement.setString(8, codecooler.getPhoneNum());
+            statement.setInt(9, codecooler.getId());
+
+            statement.executeUpdate();
+
+            connection.close();
+
+        } catch (SQLException e){
+            e.printStackTrace();
+        }
+
+    }
+
+    private int getUserIdWithLogin(User user) throws DBException {
+        ResultSet results = null;
+        PreparedStatement statement = null;
+        String query = "SELECT * FROM users WHERE login = ?;";
+        try {
+
+            connection = dbCreator.connectToDatabase();
+
+
+            statement = connection.prepareStatement(query);
+
+            statement.setString(1, user.getLogin());
+
+            results = statement.executeQuery();
+
+            results.next();
+
+            int id = results.getInt("id");
+
+            connection.close();
+
+            return id;
+
+
+
+        } catch (SQLException e) {
+            e.printStackTrace();
+            throw new DBException("Did not find user with this login SLQ Exception");
+
+        } catch (Exception e){
+            e.printStackTrace();
+            throw new DBException("Did not find user with this login");
+        }
+
+
+    }
+
 
     public int showWallet(int id) throws SQLException {
         DBCreator dbCreator = new DBCreator();
@@ -44,7 +159,6 @@ public class CodecoolerDAO implements ICodecoolerDAO {
         ResultSet rs = null;
         List<Artifact> boughtArtifacts = new ArrayList();
 
-
         try {
             stmt = con.prepareStatement("SELECT art.id, artifact_name, artifact_category, artifact_description, artifact_price\n" +
                     "FROM users_artifacts usersArt INNER JOIN artifacts art\n" +
@@ -70,11 +184,11 @@ public class CodecoolerDAO implements ICodecoolerDAO {
         }
         return boughtArtifacts;
     }
+
     //todo
     public void buyArtifactWithTeammates() {
-
-
     }
+
 
     public int showLevelOfExperience(int id) throws SQLException {
         DBCreator dbCreator = new DBCreator();
@@ -91,7 +205,6 @@ public class CodecoolerDAO implements ICodecoolerDAO {
     }
 
     private void updateMoneyAmount(int userID, int artifactID) throws SQLException{
-
         DBCreator dbCreator = new DBCreator();
         Connection connection = dbCreator.connectToDatabase();
 
@@ -118,12 +231,8 @@ public class CodecoolerDAO implements ICodecoolerDAO {
     }
 
     private boolean checkIfTransactionPossible(int userID, int artifactID) throws SQLException {
-
         int coolcoins = getCoolCoins(userID);
-
         int artifactCost = getArtifactCost(artifactID);
-
-
         System.out.println(coolcoins >= artifactCost);
         if (coolcoins >= artifactCost){return true;}
         return false;
