+++ src/main/java/dao/ICodecoolerDAO.java
@@ -3,9 +3,14 @@ package  dao;
 
 import java.sql.SQLException;
 import model.items.Artifact;
+import model.users.Codecooler;
+import model.users.User;
+
 import java.util.List;
 
 public interface ICodecoolerDAO {
+
+    void createCodecooler(User user, Codecooler codecooler);
     int showWallet(int id) throws SQLException;
 
     void buyArtifact(int userID, int artifactID) throws SQLException;
